Split the files into directories
