+++ src/render/renderer.hpp
@@ -4,15 +4,15 @@
 #include "frame.hpp"
 #include "application.hpp"
 #include "window.hpp"
-#include "vulkan/proxy.hpp"
-#include "vulkan/instance.hpp"
-#include "vulkan/compiler.hpp"
-#include "vulkan/command.hpp"
-#include "vulkan/swapchain.hpp"
-#include "vulkan/buffer.hpp"
-#include "render/vulkan/allocator.hpp"
-#include "render/vulkan/attachment.hpp"
-#include "render/vulkan/binding.hpp"
+#include "render/vulkan/setup/proxy.hpp"
+#include "render/vulkan/setup/instance.hpp"
+#include "render/vulkan/shader/compiler.hpp"
+#include "render/vulkan/command/command.hpp"
+#include "render/vulkan/setup/swapchain.hpp"
+#include "render/vulkan/buffer/buffer.hpp"
+#include "render/vulkan/setup/allocator.hpp"
+#include "render/vulkan/buffer/attachment.hpp"
+#include "render/vulkan/descriptor/vertex.hpp"
 #include "render/vulkan/pass/render.hpp"
 #include "render/vulkan/pass/pipeline.hpp"
 #include "render/vulkan/descriptor/pool.hpp"
