+++ src/render/vulkan/descriptor/layout.cpp
@@ -1,6 +1,6 @@
 
 #include "layout.hpp"
-#include "render/vulkan/device.hpp"
+#include "render/vulkan/setup/device.hpp"
 
 /*
  * DescriptorSetLayout
