+++ src/render/vulkan/descriptor/vertex.cpp
@@ -1,5 +1,5 @@
 
-#include "binding.hpp"
+#include "vertex.hpp"
 
 /*
  * BindingLayout
