+++ src/render/vulkan/pass/pipeline.cpp
@@ -1,10 +1,10 @@
 
 #include "pipeline.hpp"
-#include "render/vulkan/shader.hpp"
-#include "render/vulkan/binding.hpp"
+#include "render.hpp"
+#include "render/vulkan/shader/shader.hpp"
+#include "render/vulkan/descriptor/vertex.hpp"
 #include "render/vulkan/descriptor/layout.hpp"
-#include "render/vulkan/pass/render.hpp"
-#include "render/vulkan/device.hpp"
+#include "render/vulkan/setup/device.hpp"
 
 /*
  * GraphicsPipeline
